industrial-siemens-lib/requirements.txt
# Industrial Siemens Library - Requirements
# ========================================

# Core dependencies
python-snap7>=1.3
pymodbus>=3.5.0
asyncua>=0.60.0
cpppo>=4.3.0

# Additional industrial protocols
pyserial>=3.5
pyserial-asyncio>=0.6

# Data handling and utilities
numpy>=1.21.0
pandas>=1.3.0
structlog>=21.1.0

# Network and communication
aiohttp>=3.8.0
websockets>=10.0

# Database support (optional)
sqlalchemy>=1.4.0
asyncpg>=0.25.0

# Configuration and logging
pyyaml>=6.0
python-dotenv>=0.19.0

# Testing
pytest>=6.2.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0

# Development
black>=21.0.0
flake8>=4.0.0
mypy>=0.910

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Monitoring and diagnostics
psutil>=5.8.0
prometheus-client>=0.11.0

# Development tools
streamlit>=1.0.0

# Optional: GUI and visualization
matplotlib>=3.4.0
plotly>=5.0.0

# Optional: Database integration
sqlite3

# Optional: Cloud integration
boto3>=1.18.0
azure-storage-blob>=12.8.0
google-cloud-storage>=1.42.0 